+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Changed
 
+- Breaking! Return maps from `raycast-seq`, rather than fixtures.
+
 # 0.5.23 (2022-03-11 / 7b55d21)
 
 ## Fixed